requirements.txt
# =======================
# 📄 文档处理
# =======================
python-pptx
python-docx
PyPDF2
PyMuPDF          # fitz
pdfkit

# =======================
# 🧩 文本 / HTML 解析
# =======================
markdown2
beautifulsoup4

# =======================
# 🖼️ 图像 / OCR
# =======================
Pillow
# easyocr   # 如未来需要识别图片内文字再启用

# =======================
# 🌐 Web / 前端
# =======================
streamlit>=1.32.0   # 建议锁版本以避免兼容性问题

# =======================
# 🔒 用户认证 / 安全
# =======================
passlib==1.7.4
bcrypt==4.0.1
python-dotenv
pyperclip

# =======================
# 🧠 AI / 语言模型
# =======================
openai>=1.0.0
langdetect

# =======================
# 🗄️ 数据库 / ORM
# =======================
SQLAlchemy>=2.0.0   # ✅ 你的 routes_local 和 models.py 都在用
sqlite3             # ✅ 虽为内置库，但声明能方便云部署环境识别

# =======================
# 🧾 PDF 生成
# =======================
reportlab

# =======================
# 🧰 日志与系统监控
# =======================
loguru              # ✅ 未来 logger 模块可考虑用它增强日志格式
